refactored a few more CrudTraits to use the Settings API
