+++ src/PanelTraits/Search.php
@@ -208,7 +208,7 @@ trait Search
         }
 
         // add the details_row button to the first column
-        if ($this->get('list.details_row')) {
+        if ($this->get('list.detailsRow')) {
             $details_row_button = \View::make('crud::columns.details_row_button')
                                            ->with('crud', $this)
                                            ->with('entry', $entry)
